requirements.txt
Django>=5.1
django-unfold>=0.54.0
djangorestframework>=3.16.0
django-filter>=25.1
django-import-export>= 4.3.7
